Journalling and keystroke interception duties out of CommandManager
