+++ src/Lyrics.cpp
@@ -23,7 +23,7 @@
 #include "Project.h"
 #include "ProjectWindowBase.h"
 #include "LabelTrack.h"
-#include "commands/CommandManager.h"
+#include "MenuCreator.h"
 #include "UndoManager.h"
 #include "ViewInfo.h"
 
@@ -543,7 +543,7 @@ void LyricsPanel::OnShow(wxShowEvent &e)
 void LyricsPanel::OnKeyEvent(wxKeyEvent & event)
 {
    if (auto project = FindProjectFromWindow(this))
-      CommandManager::FilterKeyEvent(*project, event, true);
+      MenuCreator::FilterKeyEvent(*project, event, true);
    event.Skip();
 }
