+++ src/commands/CommandManager.cpp
@@ -74,27 +74,18 @@ CommandManager.  It holds the callback for one command.
 \brief List of CommandListEntry.
 
 *//******************************************************************/
-
-
-
 #include "CommandManager.h"
 
 #include "CommandContext.h"
-#include "CommandManagerWindowClasses.h"
 
 #include <wx/app.h>
 #include <wx/defs.h>
-#include <wx/evtloop.h>
 #include <wx/frame.h>
 #include <wx/hash.h>
 #include <wx/log.h>
 #include <wx/menu.h>
 
-#include "../ActiveProject.h"
 #include "BasicUI.h"
-#include "Journal.h"
-#include "JournalOutput.h"
-#include "JournalRegistry.h"
 #include "Project.h"
 #include "ProjectWindows.h"
 #include "AudacityMessageBox.h"
@@ -134,49 +125,6 @@ struct SubMenuListEntry
    TranslatableString name;
    std::unique_ptr<wxMenu> menu;
 };
-
-struct CommandListEntry
-{
-   int id;
-   CommandID name;
-   TranslatableString longLabel;
-   NormalizedKeyString key;
-   NormalizedKeyString defaultKey;
-   TranslatableString label;
-   TranslatableString labelPrefix;
-   TranslatableString labelTop;
-   wxMenu *menu;
-   CommandHandlerFinder finder;
-   CommandFunctorPointer callback;
-   CommandParameter parameter;
-
-   // type of a function that determines checkmark state
-   using CheckFn = std::function< bool(AudacityProject&) >;
-   CheckFn checkmarkFn;
-
-   bool multi;
-   int index;
-   int count;
-   bool enabled;
-   bool skipKeydown;
-   bool wantKeyup;
-   bool allowDup;
-   bool isGlobal;
-   bool isOccult;
-   bool isEffect;
-   bool excludeFromMacros;
-   CommandFlag flags;
-   bool useStrictFlags{ false };
-};
-
-NonKeystrokeInterceptingWindow::~NonKeystrokeInterceptingWindow() = default;
-
-TopLevelKeystrokeHandlingWindow::~TopLevelKeystrokeHandlingWindow() = default;
-
-bool TopLevelKeystrokeHandlingWindow::HandleCommandKeystrokes()
-{
-   return true;
-}
 
 MenuBarListEntry::MenuBarListEntry(const wxString &name_, wxMenuBar *menubar_)
    : name(name_), menubar(menubar_)
@@ -666,7 +614,7 @@ int CommandManager::NextIdentifier(int ID)
 ///WARNING: Does this conflict with the identifiers set for controls/windows?
 ///If it does, a workaround may be to keep controls below wxID_LOWEST
 ///and keep menus above wxID_HIGHEST
-CommandListEntry *CommandManager::NewIdentifier(const CommandID & nameIn,
+auto CommandManager::NewIdentifier(const CommandID & nameIn,
    const TranslatableString & label,
    wxMenu *menu,
    CommandHandlerFinder finder,
@@ -675,6 +623,7 @@ CommandListEntry *CommandManager::NewIdentifier(const CommandID & nameIn,
    int index,
    int count,
    const MenuRegistry::Options &options)
+   -> CommandListEntry*
 {
    bool excludeFromMacros =
       (options.allowInMacros == 0) ||
@@ -1107,147 +1056,6 @@ TranslatableString CommandManager::DescribeCommandsAndShortcuts(
    return result;
 }
 
-///
-///
-///
-bool CommandManager::FilterKeyEvent(AudacityProject &project, const wxKeyEvent & evt, bool permit)
-{
-   auto &cm = Get(project);
-   
-   auto pWindow = FindProjectFrame(&project);
-   CommandListEntry *entry = cm.mCommandKeyHash[KeyEventToKeyString(evt)];
-   if (entry == NULL)
-   {
-      return false;
-   }
-
-   int type = evt.GetEventType();
-
-   // Global commands aren't tied to any specific project
-   if (entry->isGlobal && type == wxEVT_KEY_DOWN)
-   {
-      // Global commands are always disabled so they do not interfere with the
-      // rest of the command handling.  But, to use the common handler, we
-      // enable them temporarily and then disable them again after handling.
-      // LL:  Why do they need to be disabled???
-      entry->enabled = false;
-      auto cleanup = valueRestorer( entry->enabled, true );
-      return cm.HandleCommandEntry(entry, NoFlagsSpecified, false, &evt);
-   }
-
-   wxWindow * pFocus = wxWindow::FindFocus();
-   wxWindow * pParent = wxGetTopLevelParent( pFocus );
-   bool validTarget = pParent == pWindow;
-   // Bug 1557.  MixerBoard should count as 'destined for project'
-   // MixerBoard IS a TopLevelWindow, and its parent is the project.
-   if( pParent && pParent->GetParent() == pWindow ){
-      if(auto keystrokeHandlingWindow = dynamic_cast< TopLevelKeystrokeHandlingWindow* >( pParent ))
-         validTarget = keystrokeHandlingWindow->HandleCommandKeystrokes();
-   }
-   validTarget = validTarget && wxEventLoop::GetActive()->IsMain();
-
-   // Any other keypresses must be destined for this project window
-   if (!permit && !validTarget )
-   {
-      return false;
-   }
-
-   auto flags = cm.GetUpdateFlags();
-
-   wxKeyEvent temp = evt;
-
-   // Possibly let wxWidgets do its normal key handling IF it is one of
-   // the standard navigation keys.
-   if((type == wxEVT_KEY_DOWN) || (type == wxEVT_KEY_UP ))
-   {
-      wxWindow * pWnd = wxWindow::FindFocus();
-      bool bIntercept =
-         pWnd && !dynamic_cast< NonKeystrokeInterceptingWindow * >( pWnd );
-
-      //wxLogDebug("Focus: %p TrackPanel: %p", pWnd, pTrackPanel );
-      // We allow the keystrokes below to be handled by wxWidgets controls IF we are
-      // in some sub window rather than in the TrackPanel itself.
-      // Otherwise they will go to our command handler and if it handles them
-      // they will NOT be available to wxWidgets.
-      if( bIntercept ){
-         switch( evt.GetKeyCode() ){
-         case WXK_LEFT:
-         case WXK_RIGHT:
-         case WXK_UP:
-         case WXK_DOWN:
-         // Don't trap WXK_SPACE (Bug 1727 - SPACE not starting/stopping playback
-         // when cursor is in a time control)
-         // case WXK_SPACE:
-         case WXK_TAB:
-         case WXK_BACK:
-         case WXK_HOME:
-         case WXK_END:
-         case WXK_RETURN:
-         case WXK_NUMPAD_ENTER:
-         case WXK_DELETE:
-         case '0':
-         case '1':
-         case '2':
-         case '3':
-         case '4':
-         case '5':
-         case '6':
-         case '7':
-         case '8':
-         case '9':
-            return false;
-         case ',':
-         case '.':
-            if (!evt.HasAnyModifiers())
-               return false;
-         }
-      }
-   }
-
-   if (type == wxEVT_KEY_DOWN)
-   {
-      if (entry->skipKeydown)
-      {
-         return true;
-      }
-      return cm.HandleCommandEntry(entry, flags, false, &temp);
-   }
-
-   if (type == wxEVT_KEY_UP && entry->wantKeyup)
-   {
-      return cm.HandleCommandEntry(entry, flags, false, &temp);
-   }
-
-   return false;
-}
-
-namespace {
-
-constexpr auto JournalCode = wxT("CM");  // for CommandManager
-
-// Register a callback for the journal
-Journal::RegisteredCommand sCommand{ JournalCode,
-[]( const wxArrayStringEx &fields )
-{
-   // Expect JournalCode and the command name.
-   // To do, perhaps, is to include some parameters.
-   bool handled = false;
-   if ( fields.size() == 2 ) {
-      if (auto project = GetActiveProject().lock()) {
-         auto pManager = &CommandManager::Get( *project );
-         auto flags = pManager->GetUpdateFlags();
-         const CommandContext context( *project );
-         auto &command = fields[1];
-         handled =
-            pManager->HandleTextualCommand( command, context, flags, false );
-      }
-   }
-   return handled;
-}
-};
-
-}
-
 /// HandleCommandEntry() takes a CommandListEntry and executes it
 /// returning true iff successful.  If you pass any flags,
 ///the command won't be executed unless the flags are compatible
@@ -1279,20 +1087,24 @@ bool CommandManager::HandleCommandEntry(const CommandListEntry * entry,
       mNiceName = {};
    }
 
-   Journal::Output({ JournalCode, entry->name.GET() });
-
    CommandContext context{ mProject, evt, entry->index, entry->parameter };
    if (pGivenContext)
       context.temporarySelection = pGivenContext->temporarySelection;
+   ExecuteCommand(context, evt, *entry);
+   return true;
+}
+
+void CommandManager::ExecuteCommand(const CommandContext &context,
+   const wxEvent *evt, const CommandListEntry &entry)
+{
    // Discriminate the union entry->callback by entry->finder
-   if (auto &finder = entry->finder) {
+   if (auto &finder = entry.finder) {
       auto &handler = finder(mProject);
-      (handler.*(entry->callback.memberFn))(context);
+      (handler.*(entry.callback.memberFn))(context);
    }
    else
-      (entry->callback.nonMemberFn)(context);
+      (entry.callback.nonMemberFn)(context);
    mLastProcessId = 0;
-   return true;
 }
 
 // Called by Contrast and Plot Spectrum Plug-ins to mark them as Last Analzers.
@@ -1872,23 +1684,3 @@ void CommandManager::TellUserWhyDisallowed(
       reason,
       helpPage);
 }
-
-#include "../KeyboardCapture.h"
-
-static KeyboardCapture::PreFilter::Scope scope1{
-[]( wxKeyEvent & ) {
-   // We must have a project since we will be working with the
-   // CommandManager, which is tied to individual projects.
-   auto project = GetActiveProject().lock();
-   return project && GetProjectFrame( *project ).IsEnabled();
-} };
-static KeyboardCapture::PostFilter::Scope scope2{
-[]( wxKeyEvent &key ) {
-   // Capture handler window didn't want it, so ask the CommandManager.
-   if (auto project = GetActiveProject().lock()) {
-      return CommandManager::FilterKeyEvent(*project, key);
-   }
-   else
-      return false;
-} };
-
